Issue #221 implement SegmentWorker and segmenting strategy phase I - strategy at start...
